Fix typo, move all trainer configs to its own file
