+++ src/battle-scene.ts
@@ -69,7 +69,8 @@ import { GameMode, GameModes, getGameMode } from "#app/game-mode";
 import FieldSpritePipeline from "#app/pipelines/field-sprite";
 import SpritePipeline from "#app/pipelines/sprite";
 import PartyExpBar from "#app/ui/party-exp-bar";
-import { trainerConfigs, TrainerSlot } from "#app/data/trainer-config";
+import { TrainerSlot } from "#app/data/trainer-config";
+import { trainerConfigs } from "./data/balance/trainers/all-trainer-configs";
 import Trainer, { TrainerVariant } from "#app/field/trainer";
 import TrainerData from "#app/system/trainer-data";
 import SoundFade from "phaser3-rex-plugins/plugins/soundfade";
