+++ src/data/dialogue.ts
@@ -1,6 +1,6 @@
 import { BattleSpec } from "#enums/battle-spec";
 import { TrainerType } from "#enums/trainer-type";
-import { trainerConfigs } from "./trainer-config";
+import { trainerConfigs } from "./balance/trainers/all-trainer-configs";
 
 export interface TrainerTypeMessages {
   encounter?: string | string[];
