+++ src/data/mystery-encounters/utils/encounter-phase-utils.ts
@@ -40,7 +40,8 @@ import { Moves } from "#enums/moves";
 import { initMoveAnim, loadMoveAnimAssets } from "#app/data/battle-anims";
 import { MysteryEncounterMode } from "#enums/mystery-encounter-mode";
 import { Status } from "#app/data/status-effect";
-import { TrainerConfig, trainerConfigs, TrainerSlot } from "#app/data/trainer-config";
+import { TrainerConfig, TrainerSlot } from "#app/data/trainer-config";
+import { trainerConfigs } from "#app/data/balance/trainers/all-trainer-configs";
 import PokemonSpecies from "#app/data/pokemon-species";
 import { Egg, IEggOptions } from "#app/data/egg";
 import { CustomPokemonData } from "#app/data/custom-pokemon-data";
