+++ src/phases/trainer-message-test-phase.ts
@@ -1,5 +1,5 @@
 import BattleScene from "#app/battle-scene";
-import { trainerConfigs } from "#app/data/trainer-config";
+import { trainerConfigs } from "#app/data/balance/trainers/all-trainer-configs";
 import { TrainerType } from "#app/enums/trainer-type";
 import { BattlePhase } from "./battle-phase";
 import { TestMessagePhase } from "./test-message-phase";
